Resolve registry renewal review findings
