+++ Packages/iOS/CmuxMobileShell/Sources/CmuxMobileShell/DeviceRegistryRefreshing.swift
@@ -31,7 +31,9 @@ public protocol DeviceRegistryRefreshing: Sendable {
     ///
     /// The same team-scoped `GET /api/devices` response that backs
     /// ``freshRoutes(forMacDeviceID:instanceTag:)``, decoded into the full two-level model
-    /// rather than narrowed to one Mac's routes. Returns a three-way outcome so
+    /// rather than narrowed to one Mac's routes. Device metadata and routes are
+    /// team-scoped; each instance's live-session summaries are filtered to the
+    /// authenticated account. Returns a three-way outcome so
     /// the caller can tell a transient failure (keep the current tree) from an
     /// auth/scope rejection (clear it). The registry is team-scoped, so a 401/403
     /// after the token/scope changed must NOT keep the previous scope's
